+++ packages/riverpod/CHANGELOG.md
@@ -1,4 +1,12 @@
-# [Unreleased minor]
+# [Unreleased major]
+
+- **Breaking** After a provider has emitted an `AsyncValue.data` or `AsyncValue.error`,
+  that provider will no-longer emit an `AsyncValue.loading`.  
+  Instead, it will re-emit the latest value, but with the property
+  `AsyncValue.isRefreshing` to true.
+
+  This allows the UI to keep showing the previous data/error when a provider
+  is being refreshed.
 
 - Adding `isLoading`, `isError`, `isData` and `asError` to `AsyncValue`.
   Those getters allow interacting with `AsyncValue` without having to rely on