+++ packages/riverpod/CHANGELOG.md
@@ -12,11 +12,12 @@
   Those getters allow interacting with `AsyncValue` without having to rely on
   pattern matching.
 
+- Fixed an issue where refreshing a `provider.future`/`provider.stream` did work properly
+
 ## 1.0.3
 
 Removed an assert preventing from overriding the same provider/family
 multiple times on a `ProviderScope`/`ProviderContainer`.
-
 
 ## 1.0.2